layout/generic/test/test_bug719523.html

<!doctype html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=719523
-->
<title>Test for Bug 719523</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=719523">Mozilla Bug 719523</a>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script>
getSelection().selectAllChildren(document);
is(document.childNodes.length, 2, "Sanity check: number of document children");
is(getSelection().focusNode, document,
   "Sanity check: document must be selected");
is(getSelection().focusOffset, 2,
   "selectAllChildren() must select all the document's children");
</script>
